# Conflicts:
#	HANAHIME/Game/resource/user/shaders/BasicShader.hlsl
#	HANAHIME/Game/resource/user/shaders/BasicShader_Huge.hlsl
#	HANAHIME/Game/src/Graphics/BasicDraw.cpp

+++ HANAHIME/Game/resource/user/shaders/BasicShader_Huge.hlsl
@@ -138,7 +138,7 @@ struct PSOutput
     float depth : SV_Target2;
     float4 normal : SV_Target3;
     float4 edgeColor : SV_Target4;
-    uint4 bright : SV_Target5;
+    float4 bright : SV_Target5;
     float4 farThanPlayerColor : SV_Target6;
 };
